loads accordians instantly.
had to and i learn thta svelte has a reactivity store.

in that ther eis inner width and and we use that to change the accordian group to be one or two, since we need them so that they close if opening another.
and to make it not flicker it bascially looks the same its just the underlaying code that chages but layout remains the same.